test: split status menu appearance coverage
